decoder: CB1 speed use u64 [1/1]
PD#SWPL-195541

Problem:
u32 overflow when calculate speed

Solution:
calculate speed use u64

Verify:
sc2

+++ drivers/frame_provider/decoder/avs2/vavs2.c
@@ -8136,9 +8136,9 @@ static unsigned long run_ready(struct vdec_s *vdec, unsigned long mask)
 	if (!dec->first_sc_checked) {
 		int size = decoder_mmu_box_sc_check(dec->mmu_box, tvp);
 		dec->first_sc_checked = 1;
-		avs2_print(dec, 0, "vavs2 cached=%d  need_size=%d speed= %d ms\n",
+		avs2_print(dec, 0, "vavs2 cached=%d  need_size=%d speed= %lld ms\n",
 			size, (dec->need_cache_size >> PAGE_SHIFT),
-			(int)(get_jiffies_64() - dec->sc_start_time) * 1000/HZ);
+			(get_jiffies_64() - dec->sc_start_time) * (1000/HZ));
 	}
 
 	if (dec->next_again_flag &&
