+++ drivers/frame_provider/decoder_v4l/avs2/vavs2.c
@@ -8136,9 +8136,9 @@ static unsigned long run_ready(struct vdec_s *vdec, unsigned long mask)
 		mmu_box = ctx->bm.mmu;
 		size = decoder_mmu_box_sc_check(mmu_box, tvp);
 		dec->first_sc_checked = 1;
-		avs2_print(dec, 0, "vavs2 cached=%d  need_size=%d speed= %d ms\n",
+		avs2_print(dec, 0, "vavs2 cached=%d  need_size=%d speed= %lld ms\n",
 			size, (dec->need_cache_size >> PAGE_SHIFT),
-			(int)(get_jiffies_64() - dec->sc_start_time) * 1000/HZ);
+			(get_jiffies_64() - dec->sc_start_time) * (1000/HZ));
 	}
 #endif
 	if (dec->next_again_flag &&
