+++ drivers/frame_provider/decoder_v4l/h264_multi/vmh264.c
@@ -12417,9 +12417,9 @@ static unsigned long run_ready(struct vdec_s *vdec, unsigned long mask)
 		int size = decoder_mmu_box_sc_check(ctx->bm.mmu, tvp);
 		hw->first_sc_checked =1;
 		dpb_print(DECODE_ID(hw), 0,
-			"vmh264 cached=%d  need_size=%d speed= %d ms\n",
+			"vmh264 cached=%d  need_size=%d speed= %lld ms\n",
 			size, (ctx->comp_info.max_size >> PAGE_SHIFT),
-			(int)(get_jiffies_64() - hw->sc_start_time) * 1000/HZ);
+			(get_jiffies_64() - hw->sc_start_time) * (1000/HZ));
 	}
 
 	if (vdec_stream_based(vdec) && (hw->init_flag == 0)
